+++ random.h
@@ -1,12 +1,12 @@
 #include <stdlib.h>
 #include <math.h>
 
-double drand() {
-	return (double)rand()/RAND_MAX;
+float drand() {
+	return (float)rand()/RAND_MAX;
 }
 
-double randn(double mu,double sigma) {
-	double r=0;
+float randn(float mu,float sigma) {
+	float r=0;
 	int n=12;
 	for(int i=0;i<n;i++) {
 		r+=drand()-0.5;
